Extend ffjord to pass additional variables to the `ode_fn`.
Currently, ffjord does not accept additional condition inputs to the `ode_fn`. Furthermore, ode solvers are not capable of computing gradients with respect to condition inputs closed over by `ode_fn`, therefore additional inputs to `ode_fn` must be exposed to the `.solve` method of ode solvers as using the `constant=` parameter. This is very useful in order to implement a conditional version of FFJORD based on the current implementation. This is in line with other bijectors (such as RealNVP) that allow additional `condition_kwargs` for its forward method.

PiperOrigin-RevId: 405569006