Use tangent spaces in the `Distribution` and `Bijector` base classes.
In the `Distribution` class we define a new `experimental_local_measure` method that emits the log probability density together with its tangent space (by default, `FullSpace`). In the `Bijector` class we define a method for obtaining density corrections. This new method will be used later by meta-bijectors and distribution subclasses like `TransformedDistribution` and `Chain` when computing `log_prob`.

In a future change we will implement a special-case that defines and uses a live-dimensions-updating method to compute the result of a projection transformation. The FillTriangular Bijector will make use of this optimization.

PiperOrigin-RevId: 406977937