Centralize generic loop utilities under internal/loop_util.
This moves `smart_for_loop` and `trace_scan` from the MCMC submodule to a more central location. I'd like to use `trace_scan` to write a pure optimization loop for `tfp.math.minimize`, but it feels weird for that to depend on the MCMC submodule, since there's nothing inherently MCMC-specific about tracing a loop.

PiperOrigin-RevId: 409045544