Simplify DeferredModule interface by combining `base_class` and `args_fn` into `build_fn`.
Previously the interface had you separately specify a base class and its deferred parameterization, but there was no good reason for this separation AFAIK. Constructing parameter dictionaries is cumbersome; it seems much more natural to just write a single function that returns a Distribution (or whatever). This interface is also more general, since it can defer an `build_fn` without knowing what type it will return.

This change is motivated by a proposal I'm working on that would use DeferredModule for trainable distributions. But it also just makes sense on its own, IMHO.

Note: this is a breaking change, which should be kosher since DeferredModule is experimental.
PiperOrigin-RevId: 409282780