Refactor ASVI to use explicit user-provided lists of substitution and surrogate rules.
This change allows the user to extend and customize ASVI by providing their own rules for specific distributions. It breaks the current logic in `_asvi_surrogate_for_distribution` apart into a list of smaller functions, one for each distribution type. The default behavior is the same, but the user now has the option to pass in custom rules.

These 'surrogate rules' are in addition to what we used to call 'substitution rules' (changed here to 'prior substitution rules' to disambiguate), which replace prior distribution parts with richer families. For consistency I use a common rule format for both rule types; this involved some changes to the substitution rule machinery.

BREAKING CHANGE: I've removed the idea of a user-writable global registry of substitution roles, including the `register_asvi_substitution_rule` helper. Instead of a registry, the user now just passes a list of rules, and we provide a reasonable default list. Relying on mutable global state was a mistake IMHO, and this helps bring the treatment of substitution rules in line with how I wanted to treat surrogate rules.

PiperOrigin-RevId: 419875323