Add `gradient_estimator` argument to `monte_carlo_variational_loss`, and support doubly-reparameterized / sticking-the-landing gradients.
This also adds a `stopped_surrogate_posterior` argument, which is required by the new gradient estimators. This can be constructed and passed automatically by `fit_surrogate_posterior_stateless` (added in a future CL), or manually by a user so inclined.

The `gradient_estimator` arg replaces the old `use_reparameterization` arg, which is now deprecated. In the interest of consistency I also merged in VIMCO as an alternative gradient estimator and deprecated the old `csiszar_vimco` function, leaving `monte_carlo_variational_loss` as a unified entry point for both single and multi-sample losses.

I chose to treat sticking-the-landing as a special case of doubly reparameterized gradients, rather than a separate estimator. The DReG paper points out that the sticking-the-landing estimator is biased in the multi-sample setting, motivating the DReG estimator, which is unbiased. I'd originally included an explicit `GradientEstimators.STICKING_THE_LANDING`, which would raise an error or perhaps compute a biased estimate if passed with `importance_sample_size > 1`. This would debatably make for clearer code in the single-sample case ('STICKING_THE_LANDING' is more self-explanatory than 'DOUBLY_REPARAMETERIZED'), but insofar as doubly-reparameterized gradients are the 'correct' generalization of sticking-the-landing to a multi-sample objective I'd rather steer people towards that then towards the original version with biased gradients.

PiperOrigin-RevId: 420178689