Add flag to GaussianProcess, such that it can retain an event_shape of [1].
Previously, if the number of (per-batch) index_points was statically
determinable and found to be 1, we would yield a Normal marginal distribution,
instead of an MVN. In some cases, users want to have a consistent event rank,
irrespective of # of index_points. We enable this behavior backward compatibly
by introducing a flag, `always_yield_multivariate_normal`, set by default to
False.

PiperOrigin-RevId: 434905157