Add experimental support for custom JAX PRNGs.
The core issue is that when custom PRNGs are enabled, the seeds become instances of PRNGKeyArrays, which act as defective np.ndarrays:

- They have no dtype field (!?)
- Most of the numpy api is not implemented for them
- They have a custom pytree registration logic
- You cannot normally access the actual integers comprising the seeds

My general strategy was to push as much of the complexity into the Numpy backend, where I had to rewrite a few of the implementations, with an eye towards general ndarray compatibility improving in the future. Most of the remaining changes deal with introspecting seeds (e.g. for seeding numpy's prngs) which is in general not allowed.

I soft-deprecated calling sanitize_seed on two-tuples, where the idea was to specify the length-2 arrays comprising seeds directly: this operation is a bit suspect for JAX, as we're violating their API encapsulation. I removed all instances of this in TFP. It's probably a good idea to completely remove this feature eventually. People should use `jax.random.PRNGKey` instead of constructing 2 element arrays manually.

Since custom PRNGs are still experimental I ran the TFP test use using the RBG PRNG manually, and added only a small amount of test coverage for the custom PRNGs in the core samplers.py library. While not everything passed, the vast majority of TFP will work with these fixes.

PiperOrigin-RevId: 435535713