Add `assertAllMeansClose` to `tfp.TestCase` for testing sampling code.
When writing a test of the form
- draw some samples
- compute the empirical mean
- assert that it's close to expected

it's useful to let the test framework itself compute that mean,
because then it can use the before-reduction samples to estimate the
test's probability of failure, and recommend how many samples to draw
to control that probability.

The new `assertAllMeansClose` method on tfp.TestCase does this,
and provides diagnostics on request.

PiperOrigin-RevId: 438389673