Use assertAllMeansClose(samples) instead of assertAllClose(reduce_mean(samples)) for IID samples.
Affected tests still pass.

Grovelling the logs from a full run indicates that we have 88
individual assertions for which the sample given by the handcoded test
seed implies a > 10% probability of chance failure, with an observed
maxiumum of 46%.  Presumably the test for which that maximum was
observed (stopping_ratio_logistic_test) is one for which the default
test seed is unlucky, but still.

PiperOrigin-RevId: 438868730