Fix JointDistributionPinned test after jax becomes more strict about inputs.
Previously jax was lax about lifting lists of arrays to arrays. It became more
strict, at least in the scipy_special.{expit,logit} implementations, which
broke a JDP test that was implicitly relying on this behavior. This change just
modifies that test to pluck out the 0th entry of a list and call jax's expit on
that entry instead of the whole list.

PiperOrigin-RevId: 441770018