Add a dynamic Cholesky option for the spike and slab sampler.
This selects the submatrix from a design matrix of active weights, and computes the Cholesky only on that submatrix. In case the design matrix is large and the problem is sparse, this may save time on O(n^3) Cholesky updates.

Due to introducing a dynamic shape into the computation, this does not work with xla compilation (and so JAX), or with batching (which would need a ragged array).

PiperOrigin-RevId: 447996071