Add a public function to convert from Gibbs posterior samples to the parameter samples used by `make_state_space_model`.
Fix the comment that assumed GibbsSamplerState could be passed to make_state_space_model previously.

Even the tests did not verify this was true since they manually mapped from the gibbs posterior to the expected parameter samples. This meant that slope_scale was never properly passed if not converted.

Also, update the anomaly detection library to use the new `model_parameter_samples_from_gibbs_samples`. Also only call it on the path that actually uses it. This removes the implicit requirement that the parameter names in GibbsSamplerStates are substrings of the STS model parameter names.

PiperOrigin-RevId: 467588955