Fix bug in BatchConcat._split_sample for scalar event shapes
Previously, the sample_shape_size for input `x` was being incorrectly computed
as

  rank(x) - shape(batch_shape)[0] - rank(event_shape)

Consider an x with sample shape [2, 3], for a batchconcat distribution with
total batch_shape [4, 5] and event_shape []. The three terms above are

  rank(x) = 2 + 2 + 0 = 4
  shape(batch_shape)[0] = (2,)[0] = 2
  rank(event_shape) = rank([]) = 1

Which computes the sample shape as 4 - 2 - 1 = 1. But it is actually 2!

The problem is that rank(shape) is always 1 no matter what, even for scalar
event shape. What we really need is ps.size(event_shape), and similarly
ps.size(batch_shape). The latter is less likely to cause problems, since batch
shape is almost always non-trivial by construction for BatchConcat.

The tests only examined the case of event_shape = [2], so the above problem
never manifested in tests. Any scalar or rank-2+ event shape would have been
incorrect.

PiperOrigin-RevId: 489347798