Avoid deprecated casting of size-1 np.ndarrays.
This used to be allowed but is now deprecated. Some logic that lies downstream
of many of our distributions' log_prob methods would invoke this behavior (in a
try/except, so it would not fail even post-deprecation, but we get an annoyting
warning all the time). This change avoids that deprecated behavior.

PiperOrigin-RevId: 574008432