Fix SNAPER step size reporting.
Internally, SNAPER absorbs part of the diagonal preconditioner inside the step
size rather than placing it entirely into the mass matrix. This makes it
difficult to compare the final step sizes to those obtained via, e.g., NUTS.
This change undoes that scaling when constructing the default trace.

Also, fix the GradientBasedTrajectoryLengthAdaptation not storing the seed in
kernel results.

PiperOrigin-RevId: 607064555