For SNAPER, use the improved gradient estimator from Riou-Durand et al.
The implementation required a bit of refactoring, since we need to be careful to only use validated (by MetropolisHastings) states for both previous and proposed states depending on direction.

Also, use harmonic mean for the docstring example, as that is best practice. The use of tfb.Exp() isn't ideal for tfd.HalfNormal(), but it gets to stay as it uncovered the issues mentioned above.

PiperOrigin-RevId: 607816872