Fix the example for tfd.Autoregressive to actually be autoregressive.
This is done by zeroing out the diagonal. The autoregressive property (with a
left-to-right order) dictates that when sampling element i, we must look only
at elements 0 ≤ k < i, which implies that the diagonal must be zero. C.f. the
'exclusive' masking for the initial layers in tfb.AutoregressiveNetwork.

Also, add a comment about the superflous iteration when sampling.

PiperOrigin-RevId: 615131666