Inference Gym: Add 64 bit support to all targets.
This is done via adding a `dtype` argument to the initializer which is then used to cast the inputs (if any) as well as the dtype for de novo arrays (like tf.zeros). This is a departure from the usual TensorFlow Probability style where the dtype is inferred from the inputs because:

- Inference Gym targets often have no inputs
- Inference Gym is not designed to do deferred array materialization like TFP is

We also depart from the typical JAX style of using the maximum precision available because we want to enable testing 32 and 64 bit implementations in the same process.

NOTE: At this time, ground truth remains always numpy arrays and always 64 bit.

Fixes #1993

PiperOrigin-RevId: 736231310